% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_differences.R
\name{mean_differences}
\alias{mean_differences}
\title{Mean differences calculation}
\usage{
mean_differences(x, y, remove_na = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{remove_na}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults are TRUE, to save time and minimize errors for users.}

\item{verbose}{Provides much more information about the flow of information between the client and server.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Numeric vector containing the differences in means.
}
\description{
Mean differences calculation
}
\examples{
mean_differences(6:8, 3:5)
}
